+++ src/math/math_util.h
@@ -10,7 +10,7 @@ float clamp(float value, float min, float max);
 void vector3f_print(vector3f v);
 void vector3d_print(vector3d v);
 void vector3i_print(vector3i v);
-int sign(float value);
+int sign(double value);
 int mod(int a, int b);
 double random_range(double min, double max);
